+++ src/QuerySerializer/Rfc3986Serializer.php
@@ -23,7 +23,8 @@ class Rfc3986Serializer implements QuerySerializerInterface
      */
     public function aggregate(array $queryParams): string
     {
-        $queryString = http_build_query(NonFiniteFloats::normalizeAll($queryParams, 'a query location value'), '', '&', PHP_QUERY_RFC3986);
+        NonFiniteFloats::assertAllFinite($queryParams, 'a query location value');
+        $queryString = http_build_query($queryParams, '', '&', PHP_QUERY_RFC3986);
 
         if ($this->removeNumericIndices) {
             $queryString = preg_replace('/%5B[0-9]+%5D/simU', '%5B%5D', $queryString);